Adds a working image viewer for user class 'admin'
Still lots of stuff to do for the image viewer. Right now the viewer uses
an iframe and javascript to change images without reloading the page, but
I started with pure php, and there's a lot of that to clean up. An upload
page is need - right now it's impossible to upload. Also some kind of
thumbnail thingy would be good.